Add new folders, separate category widgets for editing and displaying
